Organise CQRS handlers into Commands and Queries subfolders
Moves the 7 query handlers into Handlers/Queries/ and the 2 command
handlers into Handlers/Commands/, with namespaces matching the folders
(dotnet_style_namespace_match_folder). Non-CQRS handlers (Me, auth, SSE)
stay at the Handlers root. No behaviour change: handlers are discovered by
AddPostie's assembly scan, which is namespace-agnostic.
