+++ src/Wrangler.Api/Handlers/Commands/ApprovePullRequestsCommandHandler.cs
@@ -3,7 +3,7 @@ using Asm.Wrangler.Api.Requests;
 using Asm.Wrangler.Api.Services;
 using Postie.Cqrs.Commands;
 
-namespace Asm.Wrangler.Api.Handlers;
+namespace Asm.Wrangler.Api.Handlers.Commands;
 
 /// <summary>Approves and merges the specified pull requests.</summary>
 public class ApprovePullRequestsCommandHandler(IPullRequestService service)
