+++ src/Wrangler.Api/Handlers/Queries/GatesQueryHandler.cs
@@ -3,7 +3,7 @@ using Asm.Wrangler.Api.Requests;
 using Asm.Wrangler.Api.Services;
 using Postie.Cqrs.Queries;
 
-namespace Asm.Wrangler.Api.Handlers;
+namespace Asm.Wrangler.Api.Handlers.Queries;
 
 /// <summary>Lists pending deployment gates across the requested repositories.</summary>
 public class GatesQueryHandler(IGateService service)
