+++ src/Wrangler.Api/Handlers/Queries/GroupedRepositoriesQueryHandler.cs
@@ -3,7 +3,7 @@ using Asm.Wrangler.Api.Requests;
 using Asm.Wrangler.Api.Services;
 using Postie.Cqrs.Queries;
 
-namespace Asm.Wrangler.Api.Handlers;
+namespace Asm.Wrangler.Api.Handlers.Queries;
 
 /// <summary>Retrieves all accessible repositories grouped by owner, with their available workflows.</summary>
 public class GroupedRepositoriesQueryHandler(ISettingsService settingsService)
